Adjusted Rand Index

Computes the adjusted Rand index for two inputs. These inputs should be binary or categorical and of the same length. It also computes the PSNR, which is generalized here as simply \(-10 log_{10}(MSE)\). The adjusted Rand index, Jaccard Index, and Cohen's Kappa are used as a measure of the similarity of the structure of the two images. A small constant is added to the numerator and denominator of the Adjusted Rand index to ensure stability, as it is possible to have a zero denominator. The PSNR can be infinite if the error rate is 0. The Jaccard index and accuracy can account for NA values, but the Adjusted Rand and Cohen's Kappa cannot.

AdjRandIndex(x, y)

Arguments

x, y
a numeric or factor vector or image

Value

The accuracy, Jaccard index, the Adjusted Rand Index, the PSNR, and Cohen's Kappa. Note: The Jaccard index will not make sense if this is not binary.

References

Lawrence Hubert and Phipps Arabie (1985). "Comparing partitions". Journal of Classification. 2 (1): 193–218. doi: 10.1007/BF01908075

W. M. Rand (1971). "Objective criteria for the evaluation of clustering methods". Journal of the American Statistical Association. American Statistical Association. 66 (336): 846–850. doi: 10.2307/2284239

Cohen, Jacob (1960). "A coefficient of agreement for nominal scales". Educational and Psychological Measurement. 20 (1): 37–46. doi: 10.1177/001316446002000104

Examples

x <- rep(0:5, 5)
y <- c(rep(0:5, 4),rep(0,6))
AdjRandIndex(x, y)

#> Warning: Jaccard index may not make sense if more than two classes are present.

#> $Accuracy
#> [1] 0.8333333
#> 
#> $Jaccard
#> [1] 0.8
#> 
#> $AdjRand
#> [1] 0.5188621
#> 
#> $PSNR
#> [1] 7.781513
#> 
#> $Cohen
#> [1] 0.8
#>